backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3
pytest==7.4.3
requests==2.31.0